fix the contract interface
