+++ blockchain/receivers.go
@@ -41,7 +41,7 @@ func (bcr *BlockchainReactor) createAccountContract(ctx context.Context, ins str
 	AccountInfo     string `json:"account_info"`
 	ContractProgram string `json:"contract_program"`
 }) Response {
-	contract, err := bcr.accounts.CreateContract(nil, ins.AccountInfo, ins.ContractProgram)
+	contract, err := bcr.accounts.CreateContractInfo(nil, ins.AccountInfo, ins.ContractProgram)
 	if err != nil {
 		return NewErrorResponse(err)
 	}
